% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{detectOutliers}
\alias{detectOutliers}
\title{detectOutliers}
\usage{
detectOutliers(x, method = "SD", SDs = 3.5)
}
\arguments{
\item{x}{vector of values}

\item{method}{how to detect outliers: 'SD', 'Chi', 'Dixon', 'Grubbs', 'Biggest'}

\item{SDs}{if method='SD', then how many standard deviations from the mean will be considered an outlier boundary}
}
\value{
A vector of outliers from x
}
\description{
detectOutliers return values of x that are outliers.
}
\examples{
#fakeData <- data.frame(x = rnorm(4, 4, 20))
#detectOutliers(x = fakeData$x)
}

