% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMissings.R
\name{findMissings}
\alias{findMissings}
\title{findMissings}
\usage{
findMissings(data, vars, idVar = NULL)
}
\arguments{
\item{data}{data containing `vars`}

\item{vars}{list of variable names to be checked for missing}

\item{idVar}{id variable which will be printed showing which observations are missing data for vars. If NULL, then row.names() will be printed.}
}
\value{
a list of missing ids/row.names per variable listed in vars
}
\description{
findMissings provides a vector (per variable), containing the observations in the dataset.
}
\examples{
#missdata <- data.frame(id = 1:1000, C1 = sample(c(1,NA), 1000, replace=TRUE,
#                       prob=c(.8,.2)), C2 = sample(c(1,NA), 1000, replace=TRUE, prob=c(.8,.2)))
#findMissings(data = missdata, vars = c("C1", "C2"),idVar = "id")
}

