% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVars.R
\name{findVars}
\alias{findVars}
\title{findVars}
\usage{
findVars(data, str, ignore.case = TRUE, fuzzy = TRUE)
}
\arguments{
\item{data}{data.frame to search}

\item{str}{character string to search for in variable names}

\item{ignore.case}{TRUE/FALSE: Should character case be ignored when searching}

\item{fuzzy}{If TRUE, a fuzzy search method will be used}
}
\value{
A vector of matching variable names.
}
\description{
findVars searching the variable names of a data.frame for matches to a provided character string, and then returns the matching variable name(s).
}
\examples{
#fakeData <- data.frame(sex = sample(c('male', 'female'), 10, replace=T), smokng = sample(c('yes', 'no'), 10, replace=T))
#findVars(data = fakeData, str = 'smoking')
}

