% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_.R
\name{n_}
\alias{n_}
\title{n_}
\usage{
n_(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Length of `x`, ignoring NA, NaN, and Infinite values
}
\description{
n_ returns the length of a vector, excluding NA, NaN, and Infinite values.
}
\examples{
#example_vector <- sample(c(1,2,3,NA, Inf, NaN), 100, replace = T)
#length(example_vector)
#n_(example_vector)
}

