% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnv.R
\name{nnv}
\alias{nnv}
\title{nnv}
\usage{
nnv(x)
}
\arguments{
\item{x}{An R vector}
}
\value{
A vector of values that could not be coerced to numeric
}
\description{
nnv returns all values in a vector x that cannot be successfully coerced to numeric.
}
\examples{
#example_vector <- c("2", "9.123", "blue", "<10")
#nnv(example_vector)
}

