% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortDF.R
\name{sortDF}
\alias{sortDF}
\title{sortDF}
\usage{
sortDF(data, decreasing = FALSE, by = NULL)
}
\arguments{
\item{data}{data.frame to sort}

\item{decreasing}{If TRUE, will be sorted in decreasing order}

\item{by}{Variable name to sort by. If NULL, will sort by the first column}

\item{...}{other arguments to be passed to order}
}
\value{
Sorted data.frame
}
\description{
sortDF is a convenience wrapper for sorting data.frames.
}
\examples{
#example_df <- data.frame(a = sample(letters[1:3], 10, replace = T),
#                         b = rnorm(10))
#sortDF(example_df)
#sortDF(example_df, by = "b")
#sortDF(example_df, decreasing = TRUE, by = "b")
#sortDF(example_df, by=c("a", "b"))
#sortDF(example_df, by=c("b", "a"))
}

